+++ CHANGELOG.md
@@ -8,6 +8,7 @@ _These will be removed in the next major release_
 
 ## 0.29.0
 - Updated team cmdlets to workspace while retaining old names with alias https://developer.atlassian.com/cloud/bitbucket/bitbucket-api-teams-deprecation/
+- Fixed issue running `Add-BitbucketRepositoryBranchRestriction` that caused error `branch_type is only valid when branch_match_kind is branching_model
 
 ## 0.28.0
 - Added `Add-BitbucketRepositoryBranch` to create a new branch in a repo